feat(api): authorize post delete

+++ apps/api/src/controllers/posts.ts
@@ -19,14 +19,18 @@ export const getPostById = async (req: Request, res: Response) => {
 export const deletePost = async (req: Request, res: Response) => {
 	const id = Number(req.params.id)
 	try {
-		const post = await prisma.post.delete({ where: { id } })
-		return res.json({ message: "Post deleted successfully", post })
-	} catch (err) {
-		if (err instanceof PrismaClientKnownRequestError) {
-			if (err.code === "P2025") {
-				return res.status(404).json({ message: "Post not found" })
-			}
+		const post = await prisma.post.findUnique({
+			where: { id },
+		})
+		if (!post) {
+			return res.status(404).json({ message: "Post not found" })
 		}
+		if (post.authorId !== req.user!.id) {
+			return res.status(403).json({ message: "Forbidden" })
+		}
+		await prisma.post.delete({ where: { id } })
+		return res.json({ message: "Post deleted successfully", post })
+	} catch (_err) {
 		return res.status(500).json({ message: "Failed to delete post" })
 	}
 }
